Separate the logic in withErrorHandler-hoc to http-error-handler -hook
