+++ src/hoc/withErrorHandler/withErrorHandler.js
@@ -1,39 +1,19 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 import Modal from './../../components/UI/Modal/Modal';
 import Aux from './../Auxiliary/Auxiliary';
+import useHttpErrorHandler from '../../hooks/http-error-handler';
 
 const withErrorHandler = (WrappedComponent, axios) => {
     return props => {
 
-        const [error, setError] = useState(null);
-
-        //  Clear the possible existing errors.
-        const reqInterceptor = axios.interceptors.request.use(req => {
-            setError(null);
-            return req;
-        });
-        const resInterceptor = axios.interceptors.response.use(res => res, err => {
-            setError(err);
-        });
-
-        useEffect(() => {
-            return () => {
-                //  Get rid of un-needed interceptors.
-                axios.interceptors.request.eject(reqInterceptor);
-                axios.interceptors.response.eject(resInterceptor);
-            };
-        }, [reqInterceptor, resInterceptor]);
-
-        const errorConfirmedHandler = () => {
-            setError(null);
-        }
+        const [error, clearError] = useHttpErrorHandler(axios);
 
         return(
             <Aux>
                 <Modal 
                     show={error}
-                    modalClosed={errorConfirmedHandler}>
+                    modalClosed={clearError}>
                     {error ? error.message : null}
                 </Modal>
                 <WrappedComponent {...props} />
